chore: Sturctural changes
