+++ src/gitlab/GitlabClient.ts
@@ -1,6 +1,8 @@
 import { Gitlab, Pipelines } from 'gitlab'
 import { GitlabUsersDecorator } from './GitlabUsersDecorator';
-import {  GitClient, Users, Snippets, RepositoryFiles, MergeRequests, MergeRequestNotes, RepositoryOwners } from '../interfaces/GitClient';
+import {  GitClient, Users, Snippets, RepositoryFiles, 
+          MergeRequests, MergeRequestNotes, RepositoryOwners
+       } from '../interfaces/GitClient';
 import { MergeRequestDecorator } from './MergeRequestDecorator';
 import { GitlabOwners } from './GitlabOwners';
 
@@ -17,12 +19,15 @@ export class GitlabClient implements GitClient {
 
     constructor(config?) {
         this.client = new Gitlab(config)
-        this.MergeRequests = new MergeRequestDecorator(this.client.MergeRequests, this.client.Snippets, this.Users)
-        this.MergeRequestNotes = this.client.MergeRequestNotes
-        this.Pipelines = this.client.Pipelines
-        this.RepositoryFiles = this.client.RepositoryFiles
-        this.RepositoryOwners = new GitlabOwners(this.client)
-        this.Snippets = this.client.Snippets
         this.Users = new GitlabUsersDecorator(this.client.Users,this.client.Snippets)
+        this.MergeRequests = 
+            new MergeRequestDecorator(this.client.MergeRequests, this.client.Snippets, this.Users)
+
+        this.MergeRequestNotes = this.client.MergeRequestNotes
+        this.RepositoryFiles = this.client.RepositoryFiles
+        this.Pipelines = this.client.Pipelines
+        this.Snippets = this.client.Snippets
+
+        this.RepositoryOwners = new GitlabOwners(this.client)
     }
 }
