--- src/interfaces/NoteEventAttributes.ts
@@ -1,4 +1,0 @@
-export interface NoteEventAttributes {
-    note: string;
-    author_id: number;
-}
