+++ src/interfaces/structs/Job.ts
