+++ src/interfaces/structs/Pipeline.ts
