+++ src/interfaces/structs/Project.ts
@@ -1,4 +1,6 @@
 export type ProjectId = number | string;
+
 export interface Project {
     id: ProjectId
+    web_url: string
 }
