+++ src/interfaces/structs/Storage.ts
