+++ src/interfaces/structs/User.ts
